update: move to laana v0.1.2
