+++ src/laana-v0.1.2/proto/event/buddy.proto
